Deploy a prometheus-operator instance (#10)
This commit adds a new operator controller responsible for deploying an
instance of the prometheus-operator. Deploying the CRDs for the operator is optional
and is controlled by a command line flag. This allows the operator to
run seamlessly on OCP without interfering with the cluster monitoring
operator.

The controller does not create any watches on prometheus-operator
resources. It only runs periodic reconciliation and revert changes made
to these resources.