Add a CI and a release workflow (#17)
This patch adds a CI workflow and Release workflow

CI workflow is triggered on all pull-requests to main branch. It builds
the operator locally (`make operator bundle`)and also tests if the
images (operator, bundle) can be built properly.

Release workflow gets triggered when a PR is merged to main branch or if
a tag (starting with v) is pushed. The workflow builds all images -
operator, bundle and catalog and pushes those to registry.

The candidate operator can be installed by subscribing to the catalog
See: hack/olm/catalog-src.yaml

Signed-off-by: Sunil Thaha <sthaha@redhat.com>