Add an e2e suite (#18)
test(e2e): Add an e2e for the prometheus-operator controller

This commit adds an e2e test suite which can be run from the local
machine. It also adds a test which verifies that the prometheus-operator
instance which is deployed only reconciles resources with the
app.kubernetes.io/managed-by=monitoring-stack-operator label.