ci: make use of official rhobs org
This patch modifies the Makefile and the olm deployment workflow to make
use of official quay.io/rhobs org for publishing the operator, bundle
and the catalog images.

The PR also fixes an issue introduced in commit 96f52217928 which
installed an old version of `kustomize` that did not have the
`set label` command.

The Makefile is modified to use local image, i.e. no `quay.io/<org>` so
that local development does not require you to override the
`IMAGE_BASE` variable

Signed-off-by: Sunil Thaha <sthaha@redhat.com>