fix: parametrize the namespace of the prometheus operator
This commit allows running the prometheus operator in a namespace other
than default. While this was possible before, it was broken because the
service account namespace in the cluster role binding was always set as
default.

When installed through OLM, the namespace argument will be set as the current
namespace. This means that the prometheus operator will be deployed in
the same namespace as the monitoring stack operator.