ci: split workflows based on pre and post merge
Previously, On a Pull Request, the test_release workflow would execute
all jobs except the `publish-olm-development` and `create-github-release`
since it can only be done after a PR is merged. Post merge, most of the
checks expect the `e2e-tests` aren't strictly required as these are
executed prior to PR merge.

Thus, the workflow is not split into
1. Checks: which are run on every pull request
2. release: run after a PR is merged. This job decides if the release is
   going to be an OLM Development image push or a GH pre-release.

   [e2e-tests] -----> [olm-release]
                \
                 `--> [gh-pre-release]
3. candidate/stable: these olm workflows remain as is

Signed-off-by: Sunil Thaha <sthaha@redhat.com>