test: run e2e tests in the CI against an in-cluster operator
We currently spin up the operator in a goroutine when running e2e tests.
This works well for the local environment but does not verify the actual
behavior in a real Kubernetes cluster. For example, it does not take
into account whether the operator service account has all of the
required permissions.

This commit adds a flag to exclude running the operator in a goroutine
during e2e tests. It also changes the workflow in the CI to deploy the
manifests so that we run the operator in-cluster and the e2e tests
out-of-cluster. This way we can maintain a simpler workflow locally
while having more reliable tests in the CI.