test: run operators in dedicated namespace
During the e2e run, the operators run in the default namespace. This
makes it harder to do a cleanup after the test is finished. It also does
not test cases when operators are installed in the non-default namespace
which is a more realistic scenarion.

This commit changes the operators' namespace to 'operators'.