feat: deploy an instance of the grafana operator
This commit deploys an instance of the grafana operator in the same
namespace as the monitoring stack operator.

The grafana operator only has multi-namespace support for reconciling dashboards,
not for grafana instances and data sources.
Because of this, the OLM bundle will configure the grafana operator
to watch grafana resources in the same namespace and watch dashboards
from all namespaces.