refactor: move prometheus-operator assets to the OLM bundle
According to the OLM documentation, operators should not be deploying
other operators. It is the responsibility of OLM to manage the lifecycle
of operators. Further explanation is given in #49.

This commit moves the prometheus-operator assets to the OLM bundle and
allows OLM the manage all the resources related to the
prometheus-operator.