ci: run tests against olm bundle
We currently have to wait for the OLM bundle to be deployed in order to
validate whether it properly works. This lenghtens the development
feedback loop and could lead to subtle bugs with the released bundles.

This commit adds a job in the pipeline that runs the e2e tests against
an actual OLM bundle. Since the bundle image needs to be pullable, the
e2e job deploys a local registry in a kind cluster and pushes the image
to that registry. OLM can then pull the bundle and deploy it before the
e2e tests are executed.