ci: remove vanilla k8s e2e tests
We currently run e2e tests against both k8s and olm. This slows down the
CI and introduces flakes due to a high load on the shared runners.

This commit removes the plain e2e tests since the primary deployment
mechanism for the operator is going to be OLM.