feat: query Prometheus through a dedicated service
The prometheus-operated service is a statefulset governing service. It
is intended to be used by the statefulset controller for assigning a
network identity to individual pods, and is not supposed to be used for
http communication.

This commit creates a dedicated service for each Prometheus statefulset
and reconfigures the GrafanaDataSource to query it instead.