ci: capture cluster state after e2e run
Flaky tests can be hard to troubleshoot in the CI. This commit adds a
step in the pipeline which captures the cluster-state after each e2e
test run. It also adds a --retain flag which is toggled in the
pipeline to prevent deletion of the resources when a test fails.