ci: cache development tools
Previously, we have seen that downloading of tools like `oc` can fail
due to transient errors. These tools are downloaded for every CI run and
do not change often.

This PR reduces the need to download these tools over and over by making
use of  github's cache to cache all the tools downloaded by `make tools`.
The cache relies on the checksum of `.github/tools` file as key,
so any change to the `.github/tools` file will result in a new set of
tools being downloaded. The .github/tools is auto generated when running
`make tools`

Signed-off-by: Sunil Thaha <sthaha@redhat.com>