fix: fix self-scrape prometheus configuration for stacks
The prometheus instances are incorrectly configured for self-scraping
and scrape themselves twice, once through the prometheus job and once
more through the alertmanager job. The reason for this is because they
were using the app.kubernetes.io/part-of label instead of the
app.kubernetes.io/name label to identify their own service.