fix: install-plan approval logic to approve the right plan
Previously, install-plan for the desired version of Grafana Operator was
obtained by lookup the status.installPlanRef of the subscription. This
isn't reliable since the installPlanRef can point to a version higher
than desired and the reconcile loop would wait for ever without
installing Grafana.

E.g. say we wanted to install `4.0.1` and the installPlanRef refers to
`4.0.2` because by the time reconcile loop was trigger, OLM already
updated the installPlanRef to the next upgradable version - 4.0.2, the
grafana controller would wait for ever to install the operator.

This is fixed by following the steps below.
  1. Iterate  all install-plans in the monitoring-stack-operator namespace
  2. Filter the plan that matches the desired CSV version
  3. Wait for the plan to be properly initialised (i.e. status is not empty)
  4. Approve the desired install plan

Signed-off-by: Sunil Thaha <sthaha@redhat.com>