fix: establish a watch on Grafana CRs ince the CRD is present
We currently start an informer on Grafana CRs as soon as the operator
starts. This is problematic during installations because the Grafana CRD
will only become available once the grafana operator is installed. As a
result, we quite often get flakes in the CI related to deletion and
restoring of Grafana related resources.

This commit modifies the behavior of the Grafana Operator controller to
establish a watch only when the CRDs have been installed in the cluster.