fix: grafana-operator to use more optimised watches
Previously grafana-operator used `source.Kind` to watch the
resources which was far less efficient than the singleResourceInformer.

This patch makes use of `source.NewKindWithCache` which allows building
a cache/informers that watches only the `monitoring-stack-operator`
namespace and only those resources that has the label
`app.kubernetes.io/managed-by=monitoring-stack-operator`

NOTE: `msoClient` is created to make use of the cache.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>