refactor: use server side apply in mso controller (#121)
* refactor: use server side apply in mso controller

Server side apply has been in beta since kubernetes 1.16 and stable
since 1.22. It greatly reduces client complexity, since no
get-patch-update cycles are needed anymore.
Additionally server side apply allows for controllers to handle user
changes on managed resources much more easily, if that need arises in
the future.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>

* refactor: retry and skip revert test

The main refactor of this commit is to retry the test that checks if
resource changes are reverted by the operator. This mainly intends to
preempt timing issues. Additionally this test is skipped for now until
we have a solution for https://github.com/kubernetes-sigs/controller-tools/issues/676.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>