fix(olm): drop startingCSV from subscription.yaml
Remove the startingCSV line from the subscription.yaml as it's not
necessary and requires installation and updates in OLM to start at an
older release rather than just installing the latest. This can have
issues if the catalog source (index image) does not have the appropriate
starting CSV release.

(Additionally this blocked installation of
observability-operator.v0.0.11 since v0.0.10 wasn't available in the
index image, and resulted in an out of sync condition with the latest
release.)