fix: updates resources for OO and p-o
Problem: we pull the Prometheus Operator (p-o) deployment from
the upstream repo as a dependency. However this manifest sets
a very low limit to the p-o resources, this limit is easily hit
when the operator is managing multiple Prometheus instances.

Solution: remove current limits and load testing on OO and p-o
and observe the resources they consume. Establish a baseline for
both and then multiply that baseline by 3 and give some headroom
Issue https://issues.redhat.com/browse/MON-2648
Closes #166