chore: update controller-gen to v0.9.2
This mostly pulls in many fixes from controller-tools. Mainly
controller-gen now respects type marker in types and generated the CRD
manifests accordingly. This means that SSA now functions correctly with
the generated manifests. Accordingly this also reenables the
`revert-test`.
The other changes added here are mostly whitespace changes in the
manifests, the status being no longer included in the CRD manifests (see
https://github.com/kubernetes-sigs/controller-tools/issues/456 for
details) and we seem to have forgotten to regenerate some manifests
after a version bump.

The patch also updates  kustomize to 4.5.2

Fixes: https://github.com/rhobs/observability-operator/issues/142

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>