refactor: introduce common reconciler module (#185)
* refactor: introduce common reconciler module

This provides default component reconcile functions that can be used across
controllers. For more complex reconcile functions, controllers can still
create custom types.
The reconcilers are implemented as types that implement the Reconciler
interface to adhere to convention (a simple custom type would have
sufficed).
Additionally this commit renames the central controller structs to
resourceManager. They do implemenent the Reconciler interface from
controller-runtime, but we want to use the `reconciler` name for the
common module that reconciles individual resources.

* refactor: rename UpdateReconciler.controller to resourceOwner

The intention is to make the function clearer and use a less heavily
overloaded term.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>

* refactor: remove OptionalResourceReconciler

No need for a type as this can be handled by the remaining types and a
convenience function.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>

* refactor: add comments and better names to reconciler module

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>