feat: switch to file-based OLM catalogs
In order to switch to new file-based OLM catalogs, following changes
were made

  * upgrade to newer version of operator-sdk tools (operator-sdk, olm)
  * Generate existing bundles
  * Add 0.0.13 bundle to the new catalog image by
    *  creating new catalog index dockerfile
    *  adding package definition for observability-operator
    *  add 0.0.13 bundle and channels

  *  Add olm/update-channels.sh script to update olm catalog channels
  * Fix GitHub workflow to adapt to constraints introduced by the new
    file-based catalog

**NOTE:** This PR introduces a change to CI post-merge action - release.

Previously, when a commit is merged to main, the CI job used to publish
a catalog image which did not require any change to the code. With the
file-based catalog, the new bundle that is published needs to be added
to the catalog index.yaml file. Hence, the CI job release that calls the
`olm-publish` action now adds a commit on top the commit that got merged
into `main` and **pushes** that commit to the repo if a catalog image
could be published successfully.

JIRA: https://issues.redhat.com/browse/MON-2791

Signed-off-by: Sunil Thaha <sthaha@redhat.com>