feat: switch to file-based OLM catalogs
The Catalog index yaml file for the new file-based catalog will be
maintained in the olm-catalog branch. The PR introduces the changes
needed to generate new catalog index entry for new releases to the
development, candidate, and the stable channel and to push the updated
index to the olm-catalog git branch.

 To switch to new file-based OLM catalogs, following changes have been made

  * upgrade to newer version of operator-sdk tools (operator-sdk, olm)

  * Generate catalog for existing 0.0.13 bundle by
    * creating new catalog index dockerfile
    * adding package definition for observability-operator
    * adding channel definitions for all 3 channels
    see: https://github.com/rhobs/observability-operator/pull/194

  * Modify GitHub workflow to adapt to the new file-based catalog

JIRA: https://issues.redhat.com/browse/MON-2791

Signed-off-by: Sunil Thaha <sthaha@redhat.com>