build: remove outdated comment and check operator-sdk version in Makefile.tools
If operator-sdk if not at the version defined in Makefile.tools,
download the version specified. I implemented the check specifically
for operator-sdk as this can impact the release process.
We might want to either implement the version check for other tools
as well or always updated them for release targets.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>