build: publish catalog image with SHA
To enable continuous delivery of OBO to OSD/ROSA clusters via
app-interface, the OLM catalog image must be tagged with a git short
hash that represents the commit that was used for the creation of the
catalog.

This patch adds a make target `catalog-tag-sha` that tags the catalog
image with the SHA of the commit that CI bot makes when it updates the
catalog.

In addition to that the PR also update the Makefile to reflect the FBC
change that was pushed to olm-catalog branch but was missing in main.

JIRA: https://issues.redhat.com/browse/MON-2859

Signed-off-by: Sunil Thaha <sthaha@redhat.com>