feat: add resourceDiscovery status condition
The ResourceSelector field in a MonitoringStack CR can be set to nil.
This is passed on to the Prometheus CR and can be used to disable all
service discovery and thus most scraping (we add prometheus and
alertmanager scrape jobs via additionalScrapeConfigs).
When this is configured we now set a new status condition to alert users
to the fact that no service discovery is configured.

Fixes: https://issues.redhat.com/browse/MON-2864

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>