feat: add package-operator artifacts
This commit adds artifacts and make targets to build a package-operator
package image (https://github.com/package-operator/package-operator).
The package is setup to install everything into the
`observability-operator` namespace for now.
Additionally this setup currently excludes two sets of resources:
The rules for all operants are not packaged due to
https://github.com/package-operator/package-operator/issues/101.
The admission-webhook related resources are not packaged until we figure
out to add TLS certificates. Currently this is handled by OLM.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>