Package operator artifacts (#227)
* fix: add version and kind to all kustomization files

* refactor: move monitoring artifact to their own subdirectory

https://github.com/package-operator/package-operator/issues/101 prompted
me to separate the rules resources from the operator.
I think this way we're more flexible anyway. The OLM build should work
the same and including/excluding these resources should be straight
forward.

* feat: add package-operator artifacts

This commit adds artifacts and make targets to build a package-operator
package image (https://github.com/package-operator/package-operator).
The package is setup to install everything into the
`observability-operator` namespace for now.
Additionally this setup currently excludes two sets of resources:
The rules for all operants are not packaged due to
https://github.com/package-operator/package-operator/issues/101.
The admission-webhook related resources are not packaged until we figure
out to add TLS certificates. Currently this is handled by OLM.

* fix: address review comments

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>