chore: improve kind setup UX
Previously, when running hack/kind/setup.sh, the script ran its
validations such as if the tools are installed, local-registry is entry
is added to /etc/hosts etc in the middle of the run instead of the
start. This required the cluster to be deleted again and again until all
validations pass.

This commit improve the experience
  * running validations in the starting of the script
  * ensuring binaries in <project-root>/tmp/bin takes precedence
    all the time (as opposed to only when kubeclt was missing)
  * Better emojis 🔔 ✅ 🛑

Signed-off-by: Sunil Thaha <sthaha@redhat.com>