build: add package that allows an optional kubeconfig
This commit adds another package flavor and respective build and publish
targets and jobs. In this package both the observability-operator and
prometheus-operator deployments are patched so that they mount a secret
containing a kubeconfig and pass that config to the operators via
environment variables.
This package version specifically targets HyperShift setups, where the
operator needs to talk to a specific apiserver.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>