build: create branch for olm-publish action
Problem: In the make target `catalog-tag-sha` we want to use main's sha
sum to tag the catalog image. The olm-publish action however adds
commits to the local main branch before that happens, resulting in a sha
tag that doesn't correspond to main.

Solution: The olm-publish action creates a branch to add commits in.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>