chore: add OWNERS files
The OWNERS files are very simple and need to only be placed in the right
locations. Actual membership is managed through OWNERS_ALIASES.
Currently this reflects the github groups `observability-operator-contributors`
`observability-operator-maintainers`.
The current setup allows reviews and approvals by contributors to code,
but requires maintainer review for changes to build artifacts.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>