feat: deploy PrometheusRule resource for the operator (#629)
This change deploys a PrometheusRule resource with one alerting rule
firing when one controller hits more than 10% of failed reconciliations
in the last 5 minutes for more than 15 minutes.

It also removes "legacy" monitoring resources which were part of the
initial attempt to monitor the cluster observability operator but that
effort was never finalized.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>