fix: disable unmanaged Prometheus configuration (#944)
This commit adds the `--disable-unmanaged-prometheus-configuration=true`
argument to the Prometheus operator deployment. Without this change and
in case of a null resource selector, the Prometheus pods fail to become
ready because the Prometheus configuration is left empty while the
Thanos sidecar expects to find external labels. The "unmanaged
configuration" mode is an old upstream "hack" which allowed users to
provide their own custom Prometheus configuration for cases where it
couldn't be implemented with `ServiceMonitor`. The new approach is to
use the `ScrapeConfig` CRD.

When a `MonitoringStack` resource is created with a null resource
selector, the Prometheus configuration will contain no scrape
configuration and no rules. One possible use case is to deploy a
Prometheus which only ingests metrics via remote-write and/or OTLP.

Closes #932

Signed-off-by: Simon Pasquier <spasquie@redhat.com>